InfluxDB v2 API
Made some adjustments to ndj2influx
replaced some matches for single patterns with if lets
removed some unnecessary returns
restructured a few areas based on some clippy lints
Added in some crude handling for the InfluxDB v2 API.
The v2 API handling could probably be cleaned up a bit and made smoother,
but I was trying to get this done quickly and didn't want to restructure
the existing code to much.

I've added in an optional -B argument for the Bucket in InfluxDB 2. If
the bucket is passed it assumes it's v2. It then uses the database for
the "Organization" in the v2 API, and it assumes token authentication
for the v2 API. It still uses the -U command line arg, but assumes it's
an auth token instead of a username:password pair.